requirements.txt
# Core Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0

# Network Analysis
networkx>=3.1
python-igraph>=0.10.0
community>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web Interface
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Data Processing
openpyxl>=3.1.0
pyarrow>=12.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

